requirements.txt
Django==2.2.5
django-webpack-loader==0.6.0
djangorestframework==3.10.3
pytz==2019.2
django-cors-headers==3.1.1